requirements.txt
groq>=0.9.0
python-telegram-bot==21.6
python-dotenv==1.0.1
Pillow>=10.0.0
aiohttp>=3.9.0
requests==2.31.0